Added Staff features in admin panel
We have added the missing features to the Admin Panel when editing a
member of staff.
These features were available to the agent when they logged in under "My
Preferences" but were missing on the admin panel.
This can be very handy to see what a user has set for their preferences
as well as being able to change them.
We have also added the ability to edit the agent signature as HTML as
this was missing in the admin panel and the staff panel.
Added the ability to edit the department signature as HTML as this was
also missing.
Added the ability to edit the Authentication and Registration Templates
within the "Settings > Access" screen of the admin panel as this was
also missing.