Ability to disable Collaborators on ticket replies
Added the ability to enable or disable the checkbox for Collaborators to
be included in ticket replies by agents.
This can be very handy if your ticketing system receives mail from email
lists where you send an email to the list address and everyone in the
list gets a copy of the sent message.
Extremely helpful if you have agents who keep forgetting to untick the
box when replying to a ticket.