Pending Ticket Status with Ticket Auto-Close
Added a new Ticket Status called "Pending" which is a closed ticket
status.
Created a module which will allow you to set the default agent reply
status as required, if it is set to Pending all replies will have the
status set as Pending and will show in a new queue called Pending.
You can also set the Auto Close Grace Period so that tickets that have
sat in the pending queue for X amount of days are automatically closed
by the system.
New Email Template created for "Closed Tickets"
The ability has also been added to the system to send Closed Ticket
emails to customers on a per department basis.