alerts: Do not include the manager with the members
This fixes an issue, where the department manager cannot be excluded from
the new ticket alert if the members of the department are checked.
Considering there is a separate selection for the manager, this patch
adjusts the system so the manager of the department only receives the new
ticket alert if the manager checkbox is set. Otherwise, only the members of
the department other than the manager receive the alert.