orm: Use a new iteration model not requiring caching
This adds a new concept to the ORM iteration mechanisms and allows iterating
over very large queries without caching the records neither in the result
set nor in the model cache. It also implies using the mysqli_query() method
rather than the prepared statement model, as unbuffered queries with the
prepared statement system are much slower.