orm: Partially revert 8ab4432ff7
This partially reverts a change in the ORM changing the QuerySet::all()
method to return an Iterator rather than an array. I spent a while proposing
a patch to replace usage of all() with iterations; however, I cannot propose
a valid reason to abandon the previous methodology.

This reintroduces the previous behavior: calling all() will yield an array
of results from the QuerySet's iteration system.