session: Lock sessions on write to prevent thrashing
The _SESSION variable is now a special ArrayObject derivative, which will
monitor write access to itself. When writes are attempted, a lock is
acquired based on the current session_id. If multiple requests are being
handled on the same server for the same session, the subsequent requests
will be blocked if they attempt to write to the session as well.

If a wait was required when acquiring the lock, the data is re-fetched from
the storage backend and rebuilt into the array before allowing the write.
This should help guard against write thrashing for concurrent requests,
while also allowing read-only requests to proceed unblocked.

This commit also departs from the built-in session engine and handles the
session serialization in the osTicket session class directly. This is
necessary because it is not predictable when PHP will decide to commit the
session, and PHP does not support the _SESSION variable from being anything
except a plain array, which is not extensible.