forms: Fix lag updating the CDATA table
This fixes an issue where the CDATA table is updated with the OLD version of
an answer after it is updated. That is, after an answer is updated, the
CDATA table is also updated to reflect the change; however, the old content
of the answer value is used when updating the CDATA value.