Update ajax.tasks.php and ajax.tickets.php
Note: Since I'm not sure if that's all (replacing %s %s with %1$s %2$s) which is necessary to make this work, I'm sorry in advance for a potential "bad commit".

As already written on Crowdin, in German the order is wrong, so my suggestion is to make it possible to change the order by transforming "%s %s" to "%1$s and %2$s".
Format in English: Successfully assigned 4 selected tasks
Needed German: Successfully 4 selected tasks assigned