export: Fix crash on export, fix select_related
This originates from an ORM issue, where the select_related in a queryset
was not honored by the compiler. Therefore, the query would not include all
the fields requested and would require extra trips to the database for some
queries.

This is especially important for unbuffered queries, such as those used in
the export process, which be fetched in parallel with other queries.