cli: Use the option name as the item in the parseArgs() options array
Options are defined with
```
    var $options = array(
        'opt1' => array('-o', '--option', 'help'=>'...'),
    );
```
When using a particular setup, one would need to fetch the value of the
option in the module using the name `option` because it's what's used in the
long option name.

This patch fixes the system by using the given name, 'opt1', if it is
specified. This allows using a more meaningful or less verbose name
internally.