Ticket by email: reduce chances of incorrectly creating a new ticket.
One of the methods used by osTicket to find a related ticket number in an
email is by looking in the subject of the email for a ticket number, such
as [#123456]. If for some reason the subject contains two reference
numbers, such as the osTicket reference number (ie. [#543521]) and a
foreign reference number (ie. #654321), then osTicket might pickup the wrong
signature. This might result in the creation of a new ticket, while the
contents of the email should have been added to an existing ticket.
An example subject leading to this issue:

Company X #12345 Need help installing software [#987654]

Since the square opening bracket is always in the signature of the osTicket
subject anyway, this can be used to better pattern recognition. By making
this square opening bracket a requirement for ticket signature recognition,
it reduces the changes of picking up the wrong signature.

This commit adds the requirement for the square opening bracket of the
signature in the email subject.