Fix SMTP authentication when CN of certificate must match the provided hostname
I guess depending on your system settings, it might be a requirement that the
Common Name of the certificate matches the provided SMTP hostname. If not, it
will generate an error message more or less like the one below (which is being
silenced and incorrectly checked for):

stream_socket_enable_crypto(): Peer certificate CN=`*.werkonline.nu' did not
match expected CN=`83.219.92.20'

In this case the hostname was: owa.werkonline.nu
The IP address for owa.werkonline.nu is 83.219.92.20

The Socket.php file translates the hostname to an IP address, which leads to
the issue above.

This patch removes that translation from FQDN to IP address, resulting in a
successful SMTP authentication using TLS.