queue: Add options for easy addition of queues
Adds the ability to "Clone" a queue, which would be nice to create a queue
similar to the current one with a couple distinct settings, but also not a
child of the queue. Also adds the ability to add a "Sub Queue" to the current
queue, which will allow easy creation of queues which are pre-configured for
inheritance.