queue: Add conditions to whole row in queue
This allows a queue to define conditions which apply to the entire row, so
that all the columns in a row can react to a particular condition. For
instance, rows in the ticket queue for tickets due in the near future or
assigned to a particular team could be formatted a certain way.