queue: Fix several issues
This might happen if the close date was added to a queue and filtered
through a date/time filter.

Also fix warning emit when editing queue configuration

Also fix a few crashes if there is a naming mismatch on the field named in
search criteria and the current name of the field. And fix a regression
where the queue criteria would not be properly updated when saved.

If a queue inherits columns from a parent, the conditions would also be
inherited. This fix changes the association by linking the inherited
column to the child queue at run-time. This allows the queue column to
fetch the queue conditions from the child queue which inherited the
columns.

Check 'inherit columns' checkbox when cloning a queue

Fix doubled saved search listing