forms: Separate forms API from custom forms API
This appears to do nothing on the surface. However, it makes a
significant effort to completely separate the forms API from the custom
forms API.

Some background: the forms API was originally crafted for osTicket 1.8
and used to allow customization of the data associated with tickets.
Since then, the custom forms system hasn't change much; however the
forms api (class.forms.php) on which it is based, has changed
significantly. Much of the newer system is based on the forms API, such
as the forms displayed in dialogs for transfer, assignment,
quick-add administration, etc.

Because the simple forms API is still significantly intertwined with the
custom forms API (class.dynamic_forms.php), creating new field types,
and extending existing ones is problematic.

This patch changes the custom forms API to have intermediate classes
which sort-of inherit from both DynamicFormField and some subclass of
FormField, such as TextboxField. Because PHP does not support multiple
inheritance, a delegate is used, and calls to unknown methods are
delegated to the field implementation.

This removes the need for the dynamic form field and its corresponding
"impl", because the dynamic form field instance is now tightly coupled
with its corresponding simple forms type. For instance the
"DynamicTextboxField" is tightly coupled with the "TextboxField". And
the type of DynamicFormField is decided by the ORM when the instance is
created.

This separation means that the code to support both systems can be almost
completely separated. Now code for both systems has its respective places.
It also provides a much clearer approach for determining POST values from
Widgets.

FormFields nor Widgets should not access `->value` directly. Instead
`getValue` and `setValue` should be used. This practice also allows the
fields and Widget classes to intercept how the values are get and set --
probably should have been the model early on.