forms: User Form Field Edit-ability
This addresses issue 3724 where setting a form field to be non-editable by
a Users does not work. The User will still be able to edit the field when
they edit a ticket. This corrects the check on form field permissions to not
allow a User to edit the field if it’s not enabled.