oops: Fix Reset Button
This addresses issue 3670 where the reset button does nothing but return
an error. This adds a check for `data-draft-id` to the reset button. If
there is one it will delete the draft and reset the text box. If there is
not one then it will just reset the text box.
Also includes:
- A line to add the `data-draft-id` to the text box automatically, when a
draft is saved, through AJAX.