oops: Fix New Ticket Help Topic SLA
This addresses issue # where creating a ticket as an agent and
selecting a Help Topic that has an SLA plan associated with it gives the ticket the
system default SLA instead. This adds a check for the Help Topic SLA plan
and adds it to the ticket if there is one. If not then it will add whatever
SLA plan the agent selected or the system default.