lock: Require Lock For Ticket Actions
This adds locks to ticket actions so that Agents have to posses a lock in
order to do any action on the ticket. This prevents separate Agents from
doing two different actions on the same ticket. Example: Closing a Ticket
and Editing a Ticket

Lock Process:
-If Agent A does not have a lock on the ticket, the system will
generate one and will allow Agent A to perform the action.
-If Agent B already has a lock on the ticket, the system will not generate
a lock for Agent A and will prevent Agent A from doing any actions.
-If Agent A had a lock on the ticket previously but it expired the system
will try to renew it.