oops: Dept Field Set Ticket Dept
This addresses an issue where a Department Field on a ticket will not set
the Ticket’s Department upon creation. This adds a check for a Department
Field on ticket creation and if there is a value set for the field the
Ticket’s Department will be set to that value.