xss: Prevent Shared XSS Attachments
This addresses an issue where a User can upload an XSS infected
attachment to a ticket, share the file link with another person, and the
person will be able to view/download the attachment. This adds a check to
see if the attachment is apart of a ticket and if it is then only Users
associated with the ticket can view the attachment. You have to enable
the setting ‘Login Required to view attachments’ for this to work.