feature: Create Task Ticket Thread Action
This adds a Ticket Thread Action for creating a Task. This action will
popup the New Task modal and will pre-fill the body of the New Task with
the body of the Thread Entry from which the action was made from.

Example:
Thread Entry - “I will need to check on the status of your account.”
Click Create Task from that Thread’s Actions
System will show New Task popup with the body as “I will need to check on
the status of your account.”