issue: Clone Method Called On NonObject
This addresses an issue reported on the Forums where searching for
something by date will throw a PHP fatal error. (eg. Closed Date -> on ->
10/11/2017) This is due to the date value that was chosen is not a
DateTime object. The code that determines the method to search by (has a
value, before, on, etc.) for DateTime fields will sometimes clone the date
value. In order to clone the date value it has to be an object. This
updates the method that sets the date value to always return a DateTime
object for the date value.