issue: Ticket Edit Internal Note
This addresses issue 4028 where upon editing a ticket the Internal Note’s
Instructions states that it is required in order to save the changes to
the ticket. In reality this is not the case; the Internal Note is
optional. So this updates the Instructions and changes the word 'required'
to 'optional'.