issue: Helpdesk Status Help Tip
This addresses an issue mentioned on the forums where the Helpdesk Status
setting’s Help Tip is wrong. It states that if the system is in Offline
mode then Agents can still interact with the system when in reality only
Admins can interact with the system. This updates the Help Tip to inform
the Agent that only Admins will be able to access the system.