issue: Disabled & Private Help Topic
This addresses issue 3538 where upon adding a new Help Topic that is
Disabled and Private/Internal the settings do not save. Instead the Help
Topic will be Active and Public always. This happens because the way we
assign them to variables is with "double not" operators `!!`. These return
true/false values which is not what we need. We need the actual integer
for the status and type. This removes the "double not" operators to assign
the correct integer value to the variables so that the correct status and
type is saved.