deploy: Failed Bootstrap Inclusion
This addresses issue 3741 where deploying osTicket from the command line
will throw a fatal error the first time ran. The error is caused by a
failed inclusion of the `bootstrap.php` file. This is because it’s trying
to include the file from the destination folder which doesn’t exist there
yet. This updates the deploy method to use the source bootstrap file
instead of the destination bootstrap file which already exists.