issue: List & Choice Searching
This addresses issues 3703, 3493, 2625, and others where searching for
List or Choice values in a Custom Field will yield no results. This was
due to us using `FIND_IN_SET()` and `IN()` to search, which will not work
with the way the values are stored in the database. This updates the
search methods to use `REGEXP()` to search for the IDs of the List and
Choice values which will give the correct results whether it's multiple
choice or not.