i18n: FAQ Local Questions & Answers
This addresses two issues, with the first being the Featured FAQs on the
client portal are not translated when selecting another language. The
other issue is when you click the FAQ Category to see all FAQs the FAQ
names are not translated. This adds the correct `getLocal*` functions to
the FAQ templates to get the correctly translated Questions & Answers.
This also adds a new function called `getLocalTeaser` which will return
the truncated, translated FAQ answer.