issue: Drop Gmail "ltr" Attributes
This addresses an issue where the New Activity Notice Template was not
keeping the CSS styling for the %{message} variable. This was due to Gmail
adding `<div dir="ltr">` tags around some text in the body causing the CSS
styling to break. This adds RegEx to the formatter so we can remove the
`<div dir="ltr">` tags from the email body before the New Activity Notice
Template is sent preserving the CSS styling.