queue: Improve queries necessary for rendering
This removes a significant number of queries used to render the drop-down
menu for the queues.

Each queue displayed on the page previously required a database query to
determine its children. This patch changes the strategy to fetch all the
queues and organize them as a tree. The tree can then be walked as the menu
is rendered and does not require any further queries.

On my test system, it reduces the number of queries for the ticket listing
page from 56 to 46.