This commit completely removes the BCC functionality.
This means Agents can no longer choose to add a BCC User while creating or responding to a Ticket.

This also means that if a User emails in a Ticket where someone is BCC'd, the Ticket will be created without the BCC'd User(s).

If an email is sent where the Department email is BCC'd, a Ticket is created for that Department and other Users in the email are added as Collaborators.

This also changes the response options for Agents. They now have the options to:
- Reply All (User + Collaborators)
- Reply to User (only the User)
- Do not Email Reply

(Previously the options were Reply All, Reply to User, Reply to CC + User, Reply to BCC, Do not Email Reply)