Dashboard Statistics Fixes:
This commit fixes the Service Time and Response Time in the Dashboard Statistics. Now, the times will change accordingly with the date range chosen.

Service Time:
Refers to the duration of time that begins at the opening of a ticket and ends when the ticket is closed. The Service Time column measures the average Service Time per ticket, in hours.

Response Time:
Shows an average response time by an Agent, in hours, to ticket correspondence.