Agent Ticket on Behalf of User Fix:
This commit fixes an Issue that resulted from allowing Agents to add Collaborators when opening a Ticket on behalf of a User and being able to decide who will receive an email alert. If an Agent were to add an initial response to the Ticket, two email alerts would be sent out (New Ticket Notice and Response/Reply Template). Now, only the New Ticket Notice is sent out unless the Agent chooses not to send any alert.

Additionally, if the Autoresponder is set to NOT alert a User when an Agent creates a ticket on their behalf and the Agent includes an initial Agent response, the Response/Reply Template is sent out.