issue: Choices Field Export
This addresses an issue where exporting a ticket with a Custom Choices
Field does not export the Custom Choice data. This is due to the
`ChoiceField::getChoice()` function that doesn’t add the correct `id` to
the `$selection` array. This updates the function to include the `id` in
the array so we can get the correct value later on.