Issue: Dumping Tickets From User View
This commit fixes an issue with how we export tickets from a specific User. Since changes were made to how Ticket exports are done, we needed to modify the call to export the Tickets from users.php as well since we were unintentionally sending the file name in where the csv headers should've been (variable added in dumpTickets method).