Agent Ticket on Behalf of User Fix 2:
This commit fixes an issue when an Agent creates a Ticket on behalf of a User, includes an Initial Response, AND chooses 'Do Not Send Alert' for the Ticket Notice. In this instance, ONLY the Agent's initial response should be sent to the Ticket's User and Collaborators so that they will know they are included in a Ticket.

In order for this logic to work, a literal check had to be done. If the emailreply variable is set to the string literal 0 (which means do not send alert was selected), say that the alert variable in postReply is true so that the Response/Reply template will go out with only the initial Agent response. Otherwise, if emailreply is 'Alert User' or 'Alert All', the alert variable in postReply will be false and the New Ticekt Notice template will go out, showing both the user initial message and the agent initial response.