issue: Queue Export Sorting
This addresses an issue where the Queue Export does not remember your
column sorting. This was due to the `getExportableFields()` function
called in the Export Template that gets a predefined list of fields plus
some `cdata` fields with a predefined sort order (not your saved sort
order). This updates the export template to merge the `$fields` array that
contains the exportable fields in the saved sorting order and the
`getExportableFields()` result (array) that contains the same fields plus
`cdata` fields. This will return the fields in the saved sorting order
whilst still displaying possible `cdata` fields.