issue: CLI Deploy Missing Bootstrap
This addresses issue 4322 where deploying via CLI is missing
`bootstrap.php`. This is due to the `get_include_dir()` function that
tries to include a file from the destination which doesn’t exist yet,
causing a fatal error. This updates the function to use `$this->source`
instead of `$this->destination` to correctly find and include
`bootstrap.php`.