Collaborator Tickets Visibility:
This feature gives Agents the capability deciding whether or not Users should be able to see Tickets to which they are Collaborators for in their Ticket queues.

By default, Users are able to see all Tickets they take part in whether they are the User for the Ticket or a Collaborator.

If the Agent would like to disable Users seeing their Collaborator Tickets, they can configure the setting here:
Admin Panel | Tickets | Settings | Collaborator Tickets Visibility