Turn on all pages for translation purpose
The list of pages is currently filtered, that makes it impossible to translate parts of the system. This commit removes the filter for manage pages, so all pages are displayed. It also adds the other page types in the select so saving pages is possible maintaining the original page type.

Signed-off-by: Ole Kristian Ek Hornnes <ole@webapptech.no>