thread: getId On Non-Object
This addresses issue 4347 where opening a ticket without an initial message
for the User and replying to it as an agent, fatally errors. This is because
osTicket does not check if there is a last message before getting the ID
from it. So it tries to lookup the ID of a non-object therefore fatally
erroring out.