Recipients Icon + View Email Recipients for Users
This commit puts the recipient icon (for one recipient or multiple) on Thread Entries created by Users. It also stores email recipients so you can view them in the Thread Entry Actions.

For Emails:
All recipients in the email are tracked

For front end entries:
All recipients who received an alert are tracked