Reset Button Fix:
This commit makes sure that when the Reset button is hit, we only reset File Upload Field attachments for the specific form we're on.

Ex: Viewing a Ticket as an Agent and switching between 'Post Reply' and 'Post Internal Note'. If you had an image in the Post Reply file upload but switched to Post Internal Note and hit Reset, only the text/attachments on the Post Internal Note tab would be cleared.

(Improves upon #4369)