Help Topic Inline Save Fix:
This commit adds a validation for Help Topics that are saved using Inline Edit. In the event that a Ticket has a disabled Help Topic assigned to it, we want to display the name of the Help Topic, but we want to make sure that value cannot be saved.

In addition, it makes sure that if the Help Topic is disabled, it is the only disabled Help Topic that shows up in the list.