Setting - Require Help Topic to Close:
This commit adds an Admin setting that requires a Ticket to have a Help Topic before it can be closed. It can be set by going to:

Admin Panel | Settings | Tickets | Require Help Topic to Close

If enabled, Tickets will have a warning banner across the bottom of the page, a warning banner when trying to change the Ticket status to Closed, and a warning flag beside the field when editing the Ticket.

By default, this setting is NOT enabled.