issue: CDATA Phone Contains
This addresses an issue where people who do not have a phone field on the
contact information form go to search for a User in the User Directory and
the system crashes. This is due to the search query that always contains the
phone field variable. This only adds the phone variable to the search query
if the field actually exists.