Export: Make Export Fast Again
This commit addresses longstanding memory and speed issues with ticket
export especially when exporting large set of tickets.

The implementation borrows heavily on Custom Columns, which introduced the
idea of using placeholder field to parse the data from query (db) instead of
making models to do the  parsing which led to related models getting fetched
too -- and hence the memory issues.

The commit also addresses caching of choices for selection fields and such
to avoid expensive trips to database for each row!