Filter Action Send an Email Issue
This commit fixes an issue we had with adding a Filter Action to Send an Email. We had an option at the top for 'Default System Email', however, it did not have the Email ID or address for the system email and would throw an error if chosen. Now, we just retrieve the emails with their IDs and values with the Default System Email being the initial value selected in the list of options.