Retain Help Topics for Emails
If a Help Topic is disabled or archived and it is also selected as the Help Topic for an Email, make sure we do not remove that Help Topic from the Email. Instead, if that email receives a Ticket, we should just assign the ticket to the Default Help Topic. If there is no Default Help Topic, the Ticket will not have a Help Topic at all.