upgrade: Actually re-fetch the config from database
This patch fixes the issue where the configuration was not properly
reloaded during an upgrade. The configuration items were stashed in the
Config class and were not re-fetched from the ::load() method.
Additionally, this patch forces the ORM cached objects to be flushed on
each loop through the automated upgrader to ensure that cached data
which might have been modified directly in the database during the
database process is destroyed.