issue: Dupe Page Requests Fix
This addresses an issue with pull 4472 where disabling the Submit button
does not submit the Input value for the button. This affects the
installation of plugins, where the Install Path is not sent therefore the
plugin is not installed. This clones the original submit button, hides it,
then displays a dummy disabled Submit button which will submit the value and
prevent dupe posts.