feature: Public Mark As Answered/Unanswered
This feature extends the ability to mark tickets as Answered/Unanswered to
any Agent with the Role Permission to Post Reply in the Department. This
also extends the Mark As Answered option to Internal Notes and Change
Status. If you Post an Internal Note you will see a checkbox to "Mark As
Answered" same goes for changing a Ticket's Status to a status that status
is not "closed" or "resolved".