issue: Dashboard Export Period
This addresses an issue where exporting statistics via the dashboard within
a certain timeframe (eg. 07/01/2018 -> One Quarter) will always export the
statistics from the selected date up to today. This is due to the Period
option always being set to "Up to today" when the Export button is clicked
which in return exports the stats from the selected start date to today.
This adds JS to set the period the Agent selects so the Export will return
the stats from the selected date to the selected period.