Issue: Attachments on Information Fields
This commit fixes an issue where we failed to upload attachments on Information Fields. In the to_config function, we were only taking the array_values of attachments which would return an array with only the file name ([0] => 'filename') instead of an array with the file id and file name ([123] => 'filename').

Now, when we pass the keepers to keepOnlyFileIds, we have a file id that can be passed to the upload function.