merge: Set Children Statuses
This pull adds the option to set the status of all Child Tickets if setting
status of a Parent Ticket. This is helpful for situations like closing a
Parent Ticket and needing to close all Children as well. This option is only
visible when setting the status of a Ticket with Children.