profile: Reply Redirect
This pull adds a Profile Setting for Agents to choose where to redirect
after replying to ticket. The two options are Queue (redirect to queue to
work on other tickets) or Ticket (redirect back to same ticket to continue
working) with the latter being the default. The setting is configurable for
each Agent in the helpdesk.