issue: Task Response With Collaborators
This addresses a fatal error when adding a new response to a Task with
Collaborators. When we implemented the MailingList for Users/Collaborators
we forgot about the little Task Collaborators (poor little guys). This adds
a new function called `getRecipients()` to `class Thread` that utilizes
MailingList instead of UserList. This also adds two new functions to `class
Util` that retrieves the User’s `user_id`.