Option to hide Children Thread Entries
This commit gives Agents the option to hide children thread entries on parent Tickets using a checkbox on the Merge Tickets modal. If hidden, the Agent is still able to see the related children Tickets in the 'Related Tickets' tab of the Parent Ticket. The thread entries will be hidden by default.